refactor cache in telemetry
